The 'exit' version
This has two separate inputs. One that does .lower() and .strip(), and one that just does .strip(). Any time we need to verify addresses, we can't be using .lower(). So this is mostly for addresses.

Rather than sprinkling in ['exit', 'x'] everywhere, I just throw it in the inputs, and only need it in two separate spots to capture all inputs and allow exiting/quitting anywhere in the program. Nifty.